Move generated code to separate folder
