+++ .github/workflows/ci.yml
@@ -24,10 +24,4 @@ jobs:
     - run: go test -mod=mod -coverpkg="./..." -coverprofile=covprofile ./...
     - env:
         COVERALLS_TOKEN: ${{ secrets.github_token }}
-      run: >-
-        goveralls -coverprofile=covprofile -service=github -ignore="
-        confgenerator/filter/internal/errors/*,
-        confgenerator/filter/internal/lexer/*,
-        confgenerator/filter/internal/parser/*,
-        confgenerator/filter/internal/token/*,
-        confgenerator/filter/internal/util/*"
+      run: goveralls -coverprofile=covprofile -service=github -ignore="confgenerator/filter/internal/generated/*"
