+++ confgenerator/filter/internal/filter.bnf
@@ -137,7 +137,7 @@ string : '"' {_character} '"' ;
 
 import (
 	"github.com/GoogleCloudPlatform/ops-agent/confgenerator/filter/internal/ast"
-	"github.com/GoogleCloudPlatform/ops-agent/confgenerator/filter/internal/token"
+	"github.com/GoogleCloudPlatform/ops-agent/confgenerator/filter/internal/generated/token"
 )
 >>
